marketing-workflow-requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
schedule>=1.2.0
